# govconnect-dashboard/public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: GovConnect API
  description: |
    Complete API documentation for GovConnect microservices platform.
    
    ## Services
    - **Case Service** - Complaints & Service Requests management
    - **Channel Service** - WhatsApp Channel & Live Chat
    - **AI Service** - AI Orchestrator & Knowledge Base
    - **Notification Service** - Event-driven notifications
    
    ## Authentication
    - **Internal API Key**: `X-Internal-API-Key` header for service-to-service communication
    - **JWT Token**: `Authorization: Bearer <token>` for dashboard authentication
    - **Public APIs**: Endpoint publik tersedia untuk form publik (tanpa auth)
    
    ## Base URLs
    Layanan diakses langsung per service (tanpa gateway agregasi).
  version: 1.0.0
  contact:
    name: GovConnect Team
    url: https://govconnect.my.id
  license:
    name: MIT
servers:
  - url: https://case.govconnect.my.id
    description: Case Service (Production)
  - url: https://channel.govconnect.my.id
    description: Channel Service (Production)
  - url: https://ai.govconnect.my.id
    description: AI Service (Production)
...
components:
  securitySchemes:
    InternalApiKey:
      type: apiKey
      in: header
      name: X-Internal-API-Key
      description: Internal API key for service-to-service communication
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for dashboard authentication
...
paths:
...
  /internal/takeover/{wa_user_id}:
    post:
      tags:
        - Live Chat
      summary: Start takeover
      description: Start admin takeover of conversation
      operationId: startTakeover
      security:
        - InternalApiKey: []
      parameters:
        - name: wa_user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - admin_id
                - admin_name
              properties:
                admin_id:
                  type: string
                admin_name:
                  type: string
      responses:
        '200':
          description: Takeover started
    delete:
      tags:
        - Live Chat
      summary: End takeover
      operationId: endTakeover
      security:
        - InternalApiKey: []
      parameters:
        - name: wa_user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Takeover ended
...
  /internal/conversations/{wa_user_id}:
    get:
      tags:
        - Live Chat
      summary: Get conversation
      operationId: getConversation
      security:
        - InternalApiKey: []
      parameters:
        - name: wa_user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Conversation details
    delete:
      tags:
        - Live Chat
      summary: Delete conversation
      operationId: deleteConversation
      security:
        - InternalApiKey: []
      parameters:
        - name: wa_user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Conversation deleted
...